requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
bibtexparser==1.4.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
groq==0.13.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
ollama==0.3.2
openpyxl==3.1.5
pandas==2.2.2
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
pymongo==4.8.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
